talks: update repo to be deployable on App Engine again
This change moves the talk content in the content/ subdirectory
now that the default for App Engine is to use that directory per
golang.org/cl/140841.

Also update the README, app.yaml, and add a .gcloudignore since
running gcloud app deploy will add it to the file tree if it
doesn't exist.

Fixes golang/go#28080

Change-Id: Ie5dff7844d272d21ccb6db28aa1ca231f703787a
Reviewed-on: https://go-review.googlesource.com/c/141317
Reviewed-by: Andrew Gerrand <adg@golang.org>