README: clarify deploy instructions
CL 200220 made a helpful change of suggesting a convenient command
to copy the contents tools/cmd/present to the root of this repository.

This change aims to make it easier to see that the copy step is going
to be performed by one of the two commands, rather than something that
the reader must do manually first, by splitting the instructions into
two separate steps. Each step has a description and a command to run.

Update a link to use https. Also remove one of the two trailing
slashes that isn't needed, to make the command slightly shorter.

Change-Id: I69bcbd6ca37f3dfc09fd8f300e3c3e6f062aa9a9
Reviewed-on: https://go-review.googlesource.com/c/talks/+/213158
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>